Improved memory usage of music file collection.
MusicFile used to hold a reference to the metadata object generated by
mutagen. Since this isn't really needed, it is now only a local var
subject to garbage collection.

In windows holding the metadata object around caused memory to exhaust
very quickly. It might still be a problem later though.